Size mention menu to composer
