+++ src/renderer/src/components/workspace/chat/AIChatMentionPicker.tsx
@@ -13,8 +13,8 @@ import type { ProjectTreeNode } from "@shared/ipc";
 import { FileTypeIcon } from "@renderer/components/icons/FileTypeIcon";
 import {
     CHAT_COMPOSER_PICKER_MAX_HEIGHT,
-    CHAT_COMPOSER_PICKER_MAX_WIDTH,
     CHAT_COMPOSER_PICKER_MIN_WIDTH,
+    getComposerAnchoredPickerWidth,
     getViewportSafeMenuPosition,
 } from "@renderer/app/utils/menu-position";
 
@@ -47,6 +47,7 @@ interface AIChatMentionPickerProps {
 
 interface PickerPosition {
     readonly maxHeight: number;
+    readonly width: number;
     readonly x: number;
     readonly y: number;
 }
@@ -149,12 +150,11 @@ export function AIChatMentionPicker({
 
         const anchorRect = anchor.getBoundingClientRect();
         const menuRect = listRef.current?.getBoundingClientRect();
-        const width = Math.min(
-            CHAT_COMPOSER_PICKER_MAX_WIDTH,
-            Math.max(
+        const width = getComposerAnchoredPickerWidth(
+            anchorRect.width ||
+                menuRect?.width ||
                 CHAT_COMPOSER_PICKER_MIN_WIDTH,
-                Math.ceil(menuRect?.width ?? CHAT_COMPOSER_PICKER_MIN_WIDTH),
-            ),
+            window.innerWidth,
         );
         const availableHeightAbove = Math.max(0, anchorRect.top - y - 8);
         const availableHeightBelow = Math.max(
@@ -183,6 +183,7 @@ export function AIChatMentionPicker({
 
         setPosition({
             maxHeight,
+            width,
             x: safePosition.x,
             y: openAbove
                 ? Math.max(8, anchorRect.top - height - y)
@@ -249,10 +250,9 @@ export function AIChatMentionPicker({
                 left: position?.x ?? 8,
                 maxHeight:
                     position?.maxHeight ?? CHAT_COMPOSER_PICKER_MAX_HEIGHT,
-                maxWidth: CHAT_COMPOSER_PICKER_MAX_WIDTH,
-                minWidth: CHAT_COMPOSER_PICKER_MIN_WIDTH,
                 position: "fixed",
                 top: position?.y ?? 8,
+                width: position?.width ?? CHAT_COMPOSER_PICKER_MIN_WIDTH,
                 zIndex: 10010,
             }}
         >
